+++ pkgs/jni/lib/_internal.dart
@@ -30,6 +30,7 @@ export 'package:meta/meta.dart' show internal;
 export 'src/accessors.dart';
 export 'src/jni.dart' show ProtectedJniExtensions;
 export 'src/jreference.dart';
+export 'src/kotlin.dart' show coroutineSingletonsClass;
 export 'src/method_invocation.dart';
 export 'src/types.dart'
     show