+++ pkgs/jnigen/test/kotlin_test/bindings/kotlin.dart
@@ -2630,6 +2630,128 @@ class SuspendFun extends jni$_.JObject {
             .reference);
   }
 
+  static final _id_sayHelloWithoutDelay = _class.instanceMethodId(
+    r'sayHelloWithoutDelay',
+    r'(Lkotlin/coroutines/Continuation;)Ljava/lang/Object;',
+  );
+
+  static final _sayHelloWithoutDelay = jni$_.ProtectedJniExtensions.lookup<
+              jni$_.NativeFunction<
+                  jni$_.JniResult Function(
+                      jni$_.Pointer<jni$_.Void>,
+                      jni$_.JMethodIDPtr,
+                      jni$_.VarArgs<(jni$_.Pointer<jni$_.Void>,)>)>>(
+          'globalEnv_CallObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(jni$_.Pointer<jni$_.Void>,
+              jni$_.JMethodIDPtr, jni$_.Pointer<jni$_.Void>)>();
+
+  /// from: `public final java.lang.Object sayHelloWithoutDelay(kotlin.coroutines.Continuation continuation)`
+  /// The returned object must be released after use, by calling the [release] method.
+  core$_.Future<jni$_.JString> sayHelloWithoutDelay() async {
+    final $p = jni$_.ReceivePort();
+    final _$continuation = jni$_.ProtectedJniExtensions.newPortContinuation($p);
+
+    final $r = _sayHelloWithoutDelay(
+            reference.pointer,
+            _id_sayHelloWithoutDelay as jni$_.JMethodIDPtr,
+            _$continuation.pointer)
+        .object<jni$_.JObject>(const jni$_.JObjectType())
+        .reference;
+    _$continuation.release();
+    final jni$_.JReference $o;
+    if (jni$_.Jni.env.IsInstanceOf(
+      $r.pointer,
+      jni$_.coroutineSingletonsClass.reference.pointer,
+    )) {
+      $o = jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
+      $r.release();
+    } else {
+      $o = $r;
+    }
+    return const jni$_.JStringType().fromReference($o);
+  }
+
+  static final _id_failWithoutDelay = _class.instanceMethodId(
+    r'failWithoutDelay',
+    r'(Lkotlin/coroutines/Continuation;)Ljava/lang/Object;',
+  );
+
+  static final _failWithoutDelay = jni$_.ProtectedJniExtensions.lookup<
+              jni$_.NativeFunction<
+                  jni$_.JniResult Function(
+                      jni$_.Pointer<jni$_.Void>,
+                      jni$_.JMethodIDPtr,
+                      jni$_.VarArgs<(jni$_.Pointer<jni$_.Void>,)>)>>(
+          'globalEnv_CallObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(jni$_.Pointer<jni$_.Void>,
+              jni$_.JMethodIDPtr, jni$_.Pointer<jni$_.Void>)>();
+
+  /// from: `public final java.lang.Object failWithoutDelay(kotlin.coroutines.Continuation continuation)`
+  /// The returned object must be released after use, by calling the [release] method.
+  core$_.Future<jni$_.JString> failWithoutDelay() async {
+    final $p = jni$_.ReceivePort();
+    final _$continuation = jni$_.ProtectedJniExtensions.newPortContinuation($p);
+
+    final $r = _failWithoutDelay(reference.pointer,
+            _id_failWithoutDelay as jni$_.JMethodIDPtr, _$continuation.pointer)
+        .object<jni$_.JObject>(const jni$_.JObjectType())
+        .reference;
+    _$continuation.release();
+    final jni$_.JReference $o;
+    if (jni$_.Jni.env.IsInstanceOf(
+      $r.pointer,
+      jni$_.coroutineSingletonsClass.reference.pointer,
+    )) {
+      $o = jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
+      $r.release();
+    } else {
+      $o = $r;
+    }
+    return const jni$_.JStringType().fromReference($o);
+  }
+
+  static final _id_fail = _class.instanceMethodId(
+    r'fail',
+    r'(Lkotlin/coroutines/Continuation;)Ljava/lang/Object;',
+  );
+
+  static final _fail = jni$_.ProtectedJniExtensions.lookup<
+              jni$_.NativeFunction<
+                  jni$_.JniResult Function(
+                      jni$_.Pointer<jni$_.Void>,
+                      jni$_.JMethodIDPtr,
+                      jni$_.VarArgs<(jni$_.Pointer<jni$_.Void>,)>)>>(
+          'globalEnv_CallObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(jni$_.Pointer<jni$_.Void>,
+              jni$_.JMethodIDPtr, jni$_.Pointer<jni$_.Void>)>();
+
+  /// from: `public final java.lang.Object fail(kotlin.coroutines.Continuation continuation)`
+  /// The returned object must be released after use, by calling the [release] method.
+  core$_.Future<jni$_.JString> fail() async {
+    final $p = jni$_.ReceivePort();
+    final _$continuation = jni$_.ProtectedJniExtensions.newPortContinuation($p);
+
+    final $r = _fail(reference.pointer, _id_fail as jni$_.JMethodIDPtr,
+            _$continuation.pointer)
+        .object<jni$_.JObject>(const jni$_.JObjectType())
+        .reference;
+    _$continuation.release();
+    final jni$_.JReference $o;
+    if (jni$_.Jni.env.IsInstanceOf(
+      $r.pointer,
+      jni$_.coroutineSingletonsClass.reference.pointer,
+    )) {
+      $o = jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
+      $r.release();
+    } else {
+      $o = $r;
+    }
+    return const jni$_.JStringType().fromReference($o);
+  }
+
   static final _id_sayHello = _class.instanceMethodId(
     r'sayHello',
     r'(Lkotlin/coroutines/Continuation;)Ljava/lang/Object;',
@@ -2652,19 +2774,21 @@ class SuspendFun extends jni$_.JObject {
     final $p = jni$_.ReceivePort();
     final _$continuation = jni$_.ProtectedJniExtensions.newPortContinuation($p);
 
-    _sayHello(reference.pointer, _id_sayHello as jni$_.JMethodIDPtr,
+    final $r = _sayHello(reference.pointer, _id_sayHello as jni$_.JMethodIDPtr,
             _$continuation.pointer)
         .object<jni$_.JObject>(const jni$_.JObjectType())
-        .release();
+        .reference;
     _$continuation.release();
-    final $o =
-        jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
-    final $k = const jni$_.JStringType().jClass.reference;
-    if (!jni$_.Jni.env.IsInstanceOf($o.pointer, $k.pointer)) {
-      $k.release();
-      throw 'Failed';
+    final jni$_.JReference $o;
+    if (jni$_.Jni.env.IsInstanceOf(
+      $r.pointer,
+      jni$_.coroutineSingletonsClass.reference.pointer,
+    )) {
+      $o = jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
+      $r.release();
+    } else {
+      $o = $r;
     }
-    $k.release();
     return const jni$_.JStringType().fromReference($o);
   }
 
@@ -2698,19 +2822,24 @@ class SuspendFun extends jni$_.JObject {
     final $p = jni$_.ReceivePort();
     final _$continuation = jni$_.ProtectedJniExtensions.newPortContinuation($p);
     final _$string = string.reference;
-    _sayHello$1(reference.pointer, _id_sayHello$1 as jni$_.JMethodIDPtr,
-            _$string.pointer, _$continuation.pointer)
+    final $r = _sayHello$1(
+            reference.pointer,
+            _id_sayHello$1 as jni$_.JMethodIDPtr,
+            _$string.pointer,
+            _$continuation.pointer)
         .object<jni$_.JObject>(const jni$_.JObjectType())
-        .release();
+        .reference;
     _$continuation.release();
-    final $o =
-        jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
-    final $k = const jni$_.JStringType().jClass.reference;
-    if (!jni$_.Jni.env.IsInstanceOf($o.pointer, $k.pointer)) {
-      $k.release();
-      throw 'Failed';
+    final jni$_.JReference $o;
+    if (jni$_.Jni.env.IsInstanceOf(
+      $r.pointer,
+      jni$_.coroutineSingletonsClass.reference.pointer,
+    )) {
+      $o = jni$_.JGlobalReference(jni$_.JObjectPtr.fromAddress(await $p.first));
+      $r.release();
+    } else {
+      $o = $r;
     }
-    $k.release();
     return const jni$_.JStringType().fromReference($o);
   }
 }